Fix another warning
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@9270 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/commands/cmd_topic.cpp
@@ -37,7 +37,7 @@ CmdResult CommandTopic::Handle (const char* const* parameters, int pcnt, User *u
 			if (Ptr->topicset)
 			{
 				user->WriteNumeric(332, "%s %s :%s", user->nick, Ptr->name, Ptr->topic);
-				user->WriteNumeric(333, "%s %s %s %ld", user->nick, Ptr->name, Ptr->setby, Ptr->topicset);
+				user->WriteNumeric(333, "%s %s %s %lu", user->nick, Ptr->name, Ptr->setby, (unsigned long)Ptr->topicset);
 			}
 			else
 			{
